requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
pymongo==4.8.0
fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.5
textblob==0.18.0.post0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
jinja2==3.1.4
